Dev duy

+++ src/main/java/com/skincare_booking_system/entities/Services.java
@@ -26,7 +26,7 @@ public class Services {
     String description;
 
     @Column(nullable = false)
-    int price;
+    Double price;
 
     Boolean isActive;
